Move api/test_util.ts to lib/test_util.ts
